module-1/activity/poetry/chain/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core (>=0.3.76,<0.4.0)",
    "langgraph (>=0.6.7,<0.7.0)",
    "langchain-google-genai (>=2.1.12,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "app", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
